+++ tests/Unit/Private/Invoke-SetupAction.Tests.ps1
@@ -77,7 +77,7 @@ Describe 'Invoke-SetupAction' -Tag 'Private' {
         @{
             MockParameterSetName = 'PrepareImage'
             # cSpell: disable-next
-            MockExpectedParameters = '-PrepareImage -AcceptLicensingTerms -MediaPath <string> -InstanceName <string> -Features <string[]> -InstanceId <string> [-IAcknowledgeEntCalLimits] [-Enu] [-UpdateEnabled] [-UpdateSource <string>] [-InstallSharedDir <string>] [-InstanceDir <string>] [-PBEngSvcAccount <string>] [-PBEngSvcPassword <securestring>] [-PBEngSvcStartupType <string>] [-PBStartPortRange <ushort>] [-PBEndPortRange <ushort>] [-PBScaleOut] [-Timeout <uint>] [-Force] [-WhatIf] [-Confirm] [<CommonParameters>]'
+            MockExpectedParameters = '-PrepareImage -AcceptLicensingTerms -MediaPath <string> -Features <string[]> -InstanceId <string> [-IAcknowledgeEntCalLimits] [-Enu] [-UpdateEnabled] [-UpdateSource <string>] [-InstallSharedDir <string>] [-InstanceDir <string>] [-PBEngSvcAccount <string>] [-PBEngSvcPassword <securestring>] [-PBEngSvcStartupType <string>] [-PBStartPortRange <ushort>] [-PBEndPortRange <ushort>] [-PBScaleOut] [-Timeout <uint>] [-Force] [-WhatIf] [-Confirm] [<CommonParameters>]'
         }
         @{
             MockParameterSetName = 'CompleteImage'
@@ -3188,7 +3188,6 @@ Describe 'Invoke-SetupAction' -Tag 'Private' {
                         PrepareImage = $true
                         AcceptLicensingTerms = $true
                         MediaPath = '\SqlMedia'
-                        InstanceName = 'INSTANCE'
                         # Intentionally using both upper- and lower-case.
                         Features = 'SqlEngine'
                         InstanceId = 'Instance'
@@ -3205,7 +3204,6 @@ Describe 'Invoke-SetupAction' -Tag 'Private' {
                             $ArgumentList | Should -MatchExactly '\/ACTION=PrepareImage'
                             $ArgumentList | Should -MatchExactly '\/IACCEPTSQLSERVERLICENSETERMS' # cspell: disable-line
                             $ArgumentList | Should -MatchExactly '\/FEATURES=SQLENGINE'
-                            $ArgumentList | Should -MatchExactly '\/INSTANCENAME="INSTANCE"' # cspell: disable-line
                             $ArgumentList | Should -MatchExactly '\/INSTANCEID="Instance"' # cspell: disable-line
 
                             # Return $true if none of the above throw.
@@ -3224,7 +3222,6 @@ Describe 'Invoke-SetupAction' -Tag 'Private' {
                             $ArgumentList | Should -MatchExactly '\/ACTION=PrepareImage'
                             $ArgumentList | Should -MatchExactly '\/IACCEPTSQLSERVERLICENSETERMS' # cspell: disable-line
                             $ArgumentList | Should -MatchExactly '\/FEATURES=SQLENGINE'
-                            $ArgumentList | Should -MatchExactly '\/INSTANCENAME="INSTANCE"' # cspell: disable-line
                             $ArgumentList | Should -MatchExactly '\/INSTANCEID="Instance"' # cspell: disable-line
 
                             # Return $true if none of the above throw.
@@ -3307,7 +3304,6 @@ Describe 'Invoke-SetupAction' -Tag 'Private' {
                         PrepareImage = $true
                         AcceptLicensingTerms = $true
                         MediaPath = '\SqlMedia'
-                        InstanceName = 'INSTANCE'
                         # Intentionally using both upper- and lower-case.
                         Features = 'SqlEngine'
                         InstanceId = 'Instance'