+++ CHANGELOG.md
@@ -42,7 +42,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Fix duplication of SQL Server Browser Firewall Rule when deploying
     Analysis Services feature ([issue #1942](https://github.com/dsccommunity/SqlServerDsc/issues/1942)).
 - SqlLogin
-  - Attempting to disable and already disabled login throws an error ([issue #1952](https://github.com/dsccommunity/SqlServerDsc/issues/1952))
+  - Attempting to disable and already disabled login throws an error ([issue #1952](https://github.com/dsccommunity/SqlServerDsc/issues/1952)).
+- `Install-SqlDscServer`
+  - Now the parameter `InstanceName` can no longer be specified (as per
+    the SQL Server documentation) for the setup action `PrepareImage`
+    ([issue #1960](https://github.com/dsccommunity/SqlServerDsc/issues/1960)).
 
 ## [16.3.1] - 2023-05-06